Reduced the size of the playback buffer for the following reason:
When working on consecutive frames and using play/pause a lot, it sometimes
seems as if the corresponding sound of the frames were overlapping.

I think this might be caused by a combination of SDL's playback and the
PlaybackThread's callback function.

So, when SDL wants more samples, it calls the callback function and provides it
with a pointer to some arbitrary memory which then has to be filled with new
samples.  Also, SDL obviously uses additional memory internally to provide the
audio device with data. Clearly, this memory and the memory pointed to by the
callback function can never be the same.

SDL probably uses an internal pointer that points to it's current internal
playback position. Thus, if the user stops playback, SDL stops it's own
playback, yet leaves the beforementioned pointer unchanged. This is why there
may be a few more samples left for playback inside SDL's internal sample
memory.  As soon as the user starts playback again, SDL first plays back the
remaining samples and then the new ones.

Summa summarum this means that in order to avoid the apparent overlap one can
either reduced the size of the playback buffer or only start playback once and
never stop providing SDL with new samples, may it be real samples or silence.
I prefer the first of the two possible solutions. ;-)